requirements.txt
torch
torchvision
matplotlib
pytz
pillow
pandas
tensorboardX
pyyaml
numpy